- Hovering for striped tables is not shown. - TableView background is not correct if 'stripe table' is disabled for keyed tables. - Disabling 'wrap string' option also disables 'Stripe table' option. - Key column symbol is not shown for keyed table but dictionary.
