+++ version.properties
@@ -1,1 +1,1 @@
-pluginVersion=0.22.0
+pluginVersion=0.23.0
